+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 * made class MaterialSwitchTrack internal.
 * made class MaterialProperties internal.
+* made MaterialSwitchProperty and all subclasses internal.
 
 * ## [0.9.2-preview] - 2023-05-11